Fix build failure and other improvements
- moved quarkus snippet tests to quarkus.ls
- generated microprofile.ls test jar so test classes can be reused in
quarkus tests
- fixed failing tests when running on Java > 12
- added buildAll scripts for convenience
