+++ README.md
@@ -4,3 +4,5 @@ This project contains:
 
  * [MicroProfile JDT LS Extensions](https://github.com/redhat-developer/quarkus-ls/tree/master/microprofile.jdt)
  * [MicroProfile Language Server](https://github.com/redhat-developer/quarkus-ls/tree/master/microprofile.ls)
+
+ You can build all projects at once by running the `buildAll.sh` script (`buildAll.bat` on Windows).
