+++ buildAll.bat
@@ -1,0 +1,4 @@
+cd microprofile.jdt && .\mvnw.cmd clean verify ^
+&& cd ..\microprofile.ls\com.redhat.microprofile.ls && .\mvnw.cmd clean install ^
+&& cd ..\..\quarkus.ls.ext\com.redhat.quarkus.ls && .\mvnw.cmd clean verify ^
+&& cd ..\..
