+++ quarkus.ls.ext/com.redhat.quarkus.ls/pom.xml
@@ -154,6 +154,13 @@
 			<scope>provided</scope>
 		</dependency>
 		<dependency>
+			<groupId>com.redhat.microprofile</groupId>
+			<artifactId>com.redhat.microprofile.ls</artifactId>
+			<classifier>tests</classifier>
+			<version>${microprofile.ls.version}</version>
+			<scope>test</scope>
+		</dependency>
+		<dependency>
 			<groupId>junit</groupId>
 			<artifactId>junit</artifactId>
 			<version>4.12</version>
